+++ resources/views/tuition_waiver/index.blade.php
@@ -3,6 +3,7 @@
 @extends('layouts.app')
 
 @section('content')
+<?php $allowChanges = Auth::user()->role->name === 'Director' ?>
 <div class="container">
     <div class="row">
 
@@ -14,26 +15,33 @@
         <div class="col-md-7">
             <div class="panel-group">
 
+                <div class="btn-group">
+	            	<a class="btn btn-default" id="expand_all">Expand All</a>
+	            	<a class="btn btn-default" id="collapse_all">Collapse All</a>
+	            </div>
+
                 @foreach ($waivers as $waiver)
                     <div class="panel panel-primary">
                         <div class="panel-heading clearfix">
                             <div class="panel-title pull-left" style="padding-top: 4px;">
-                                <a data-toggle="collapse" href="#collapse{{ $waiver->id }}">
+                                <a data-toggle="collapse" href="#collapse_outer{{ $waiver->id }}">
                                     {{ $waiver->description }}
                                 </a>
                             </div>
-                            <div class="btn-group pull-right">
-                                {!! Form::open(['method' => 'DELETE', 'route' => ['tuition_waiver.delete', $waiver], 'class' => 'form-horizontal', 'onsubmit' => 'return ConfirmDelete()']) !!}
-                                    <a href="{{ route('tuition_waiver.update', $waiver) }}" class="btn btn-default btn-sm" data-toggle="tooltip" title="Edit">
-                                        <span class="glyphicon glyphicon-edit"></span>
-                                    </a>
-                                    <button type="submit" class="btn btn-default btn-sm" data-toggle="tooltip" title="Delete" >
-                                        <span class="glyphicon glyphicon-trash"></span>
-                                    </button>
-                                {!! Form::close() !!}
-                            </div>
+                            @if ($allowChanges)
+                                <div class="btn-group pull-right">
+                                    {!! Form::open(['method' => 'DELETE', 'route' => ['tuition_waiver.delete', $waiver], 'class' => 'form-horizontal', 'onsubmit' => 'return ConfirmDelete()']) !!}
+                                        <a href="{{ route('tuition_waiver.update', $waiver) }}" class="btn btn-default btn-sm" data-toggle="tooltip" title="Edit">
+                                            <span class="glyphicon glyphicon-edit"></span>
+                                        </a>
+                                        <button type="submit" class="btn btn-default btn-sm" data-toggle="tooltip" title="Delete" >
+                                            <span class="glyphicon glyphicon-trash"></span>
+                                        </button>
+                                    {!! Form::close() !!}
+                                </div>
+                            @endif
                         </div>
-                        <div id="collapse{{ $waiver->id }}" class="panel-collapse collapse">
+                        <div id="collapse_outer{{ $waiver->id }}" class="panel-collapse collapse">
                             <ul class="list-group">
                                 <li class="list-group-item">Date Received: {{ $waiver->date_received }}</li>
                                 <li class="list-group-item">Amount Received: {{ $waiver->amount_received }}</li>
@@ -50,12 +58,14 @@
             </div> <!-- /.panel-group -->
         </div> <!-- /.col-md-7 -->
 
-        <!-- Affixed side nav for 'Add a Tuition Waiver' button -->
-        <nav class="col-md-2">
-        	<div data-spy="affix" data-offset-top="-1">
-    			<a href="{{ url('/waiver/add') }}" class="btn btn-success btn-lg">Add a Tuition Waiver</a>
-        	</div>
-    	</nav> <!-- /.col-md-2 -->
+        @if ($allowChanges)
+            <!-- Affixed side nav for 'Add a Tuition Waiver' button -->
+            <nav class="col-md-2">
+            	<div data-spy="affix" data-offset-top="-1">
+        			<a href="{{ url('/waiver/add') }}" class="btn btn-success btn-lg">Add a Tuition Waiver</a>
+            	</div>
+        	</nav> <!-- /.col-md-2 -->
+    	@endif
 
     </div>
 </div>
